Add snake sample This sample need slightly modified version of sleep, to be able play okayish. Proper C way is to have `nanosleep` which require interop of structs between C and Cesium runtime which is not there yet. I would like to temporary have non-standard `sleep` function, because it allow us to have tech demo

+++ Cesium.Compiler/stdlib/unistd.h
@@ -1,2 +1,2 @@
 __cli_import("Cesium.Runtime.UniStdFunctions::Sleep")
-int sleep(unsigned int);
+int sleep(double);
