+++ Cesium.Runtime/UniStdFunctions.cs
@@ -2,7 +2,7 @@ namespace Cesium.Runtime;
 
 public static class UniStdFunctions
 {
-    public static int Sleep(uint duration)
+    public static int Sleep(double duration)
     {
         Thread.Sleep((int)(duration * 1000));
         return 0;
